+++ NEWS.md
@@ -9,3 +9,5 @@
 * 26-06-2020 0.0.0.1800 Added a sps data-set to illustrate many of the internal functions.
 
 * 03-07-2020 0.0.0.1750 Added categoryplot and other utility functions
+
+* 09-07-202 0.0.0.1700 added examinedata and examinevar to facilitate the examination and characterization of fishery independent data. This is aimed at simplifying data selection prior to standardization.